requirements.txt
# Core dependencies
matplotlib>=2.1.0
pandas>= 0.22.0
scikit-learn>=0.19.1
scipy>=1.0.0
svgwrite>=1.1.12
tensorflow==2.12.0
tensorflow-probability==0.20.1
openpyxl

# Flask dependencies
flask>=2.3.2
Flask-Login>=0.6.2
Flask-SQLAlchemy>=3.0.0
Flask-WTF>=1.1.1
WTForms>=3.0.1
email-validator>=2.0.0
Flask-Compress>=1.13
Flask-Caching>=2.0.0
Flask-Limiter>=3.3.0
Flask-Minify>=0.42
flask-assets>=2.0
rjsmin
rcssmin
alembic>=1.13.0

# Note: For documentation dependencies, see docs/requirements.txt